File Uploads connection to Audit. Refactoring.
