+++ backend/audit/exceptions.py
@@ -8,3 +8,10 @@ class SessionWarningException(Exception):
     def __init__(self, message="Your session is about to expire."):
         self.message = message
         super().__init__(self.message)
+
+
+class LateChangeError(Exception):
+    """
+    Exception covering attempts to change submissions that don't have the in_progress
+    status.
+    """
