+++ backend/audit/models/__init__.py
@@ -5,32 +5,33 @@ from .access import (
 )
 from .deleted_access import DeletedAccess
 from .access_roles import ACCESS_ROLES
-from .models import (
+from .files import (
     ExcelFile,
+    excel_file_path,
+    SingleAuditReportFile,
+    single_audit_report_path,
+)
+from .models import (
     GeneralInformationMixin,
-    LateChangeError,
     SingleAuditChecklist,
     SingleAuditChecklistManager,
-    SingleAuditReportFile,
-    SacValidationWaiver,
-    UeiValidationWaiver,
     User,
-    excel_file_path,
     generate_sac_report_id,
-    single_audit_report_path,
 )
 from .audit import Audit
 from .history import History
 from .submission_event import SubmissionEvent
+from .waivers import SacValidationWaiver, UeiValidationWaiver
+from ..exceptions import LateChangeError
 
 # In case we want to iterate through all the models for some reason:
 _models = [
     Access,
     Audit,
     DeletedAccess,
+    History,
     ExcelFile,
     GeneralInformationMixin,
-    History,
     SubmissionEvent,
     LateChangeError,
     SingleAuditChecklist,
