+++ backend/dissemination/remove_singleauditreport_pdf.py
@@ -1,6 +1,6 @@
 import logging
 
-from audit.models.models import SingleAuditReportFile
+from audit.models.files import SingleAuditReportFile
 
 from dissemination.remove_workbook_artifacts import delete_files_in_bulk
